refactor: 테스트 코드 위치 테스트 클래스에 동기화
